requirements.txt
scikit-learn>=0.13.0
sklearn-pandas==0.0.10
scikits.audiolab==0.11.0
scipy>=0.10.0
numpy>=1.0.0
Flask==0.10.1
Flask-RESTful==0.3.3
Flask-SQLAlchemy==2.0
pandas>=0.15.0
-e git+git://github.com/jameslyons/python_speech_features.git#egg=speech-features
